requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastar==0.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.3
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
schedule==1.2.2
sentry-sdk==2.56.0
shellingham==1.5.4
starlette==1.0.0
telebot==0.0.5
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.2.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
